add nftIssuedEvent in video mappings
