+++ query-node/mappings/src/content/utils.ts
@@ -28,6 +28,11 @@ import {
   DataObjectTypeChannelCoverPhoto,
   DataObjectTypeVideoMedia,
   DataObjectTypeVideoThumbnail,
+  ContentActorLead,
+  ContentActorCurator,
+  ContentActorMember,
+  ContentActor as ContentActorVariant,
+  Curator,
 } from 'query-node/dist/model'
 // Joystream types
 import { ContentActor, StorageAssets } from '@joystream/types/augment'
@@ -294,6 +299,50 @@ export async function convertContentActorToChannelOrNftOwner(
   }
 
   // TODO: contentActor.isLead
+
+  logger.error('Not implemented ContentActor type', { contentActor: contentActor.toString() })
+  throw new Error('Not-implemented ContentActor type used')
+}
+
+export async function convertContentActor(
+  store: DatabaseManager,
+  contentActor: ContentActor
+): Promise<typeof ContentActorVariant> {
+  if (contentActor.isMember) {
+    const memberId = contentActor.asMember.toNumber()
+    const member = await store.get(Membership, { where: { id: memberId.toString() } as FindConditions<Membership> })
+
+    // ensure member exists
+    if (!member) {
+      return inconsistentState(`Actor is non-existing member`, memberId)
+    }
+
+    const result = new ContentActorMember()
+    result.member = member
+
+    return result
+  }
+
+  if (contentActor.isCurator) {
+    const curatorId = contentActor.asCurator[1].toNumber()
+    const curator = await store.get(Curator, {
+      where: { id: curatorId.toString() } as FindConditions<Curator>,
+    })
+
+    // ensure curator group exists
+    if (!curator) {
+      return inconsistentState('Actor is non-existing curator group', curatorId)
+    }
+
+    const result = new ContentActorCurator()
+    result.curator = curator
+
+    return result
+  }
+
+  if (contentActor.isLead) {
+    return new ContentActorLead()
+  }
 
   logger.error('Not implemented ContentActor type', { contentActor: contentActor.toString() })
   throw new Error('Not-implemented ContentActor type used')
